Enables anonymous proving. Moves background tasks out of ui module
